backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
celery==5.3.6
redis==5.0.1
groq==0.4.2
youtube-transcript-api==0.6.2
google-api-python-client==2.111.0
sentence-transformers==2.3.1
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
numpy==1.26.2

